requirements.txt
Flask==3.1.0
flask-cors==5.0.1
firebase-admin==6.8.0
python-dotenv==1.0.1
flasgger==0.9.7.1
PyJWT==2.10.1
gunicorn==23.0.0